Removed support for v 0.6
It was too difficult to keep support for v 0.6.

For anyone using v 0.6, the previous version should be used. The bug in
polynomial fit was there for so long that anyone that experienced it
would not be using this version. Therefore I think it won't be a problem
to just drop version 0.6 support.