Added documentation and tests for the Gauss-Newton algorithm
There are two interfaces for the Gauss-Newton algorithm. In
one case, the fitting function is of the type y  = f(x, a) where a is
an array. In the second interface, a generic least squares interface,
the function is of the form f(x, a) = 0 where, now, both x and a are vectors.