Hopefully corrected bug on `linear_rational_fit`
This was a bad bug! Long ago, to compute a linear fit, ther QR was directly used. This worked most of the time but sometimes ita failed spectacularly. This was brought up by a user who proposed a solution. This solution was implemented in the function `fit_linear_model`. Since the QR naïve solution was so simple, I could just use the code A\y to compute the least squares. I forgot to replace this call for the correct one. Unit testing on several platforms really helps!