Added compatibility with Polynomials 4.0
The issue was that the constructor of Polynomials checks for nonzero polynomial coefficients. For nonlinear curve fitting of rational polynomials this was a problem. Now `RationalPolynomial` just stores the coefficients directly and doesn't use `Polynomials` at all.

The `Polynomials` package has grown a lot! I have to check if it has rational polynomials already.